+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Registration module, Editing (manual): control point color corresponds to modification status (red: modified, green: not modified) and transparency corresponds to frame range selection status (opaque: selected, semi-transparent: not selected).
 * Metadata viewer: add image data type (dtype).
 * Ground truth module: input images can have T or Z axis (but no C axis), add option to export cellpose training set (with optional out-of-focus images), add option to start with an existing mask, generate logfile, save metadata.
+* File organization module: reorganize the GUI (only one "Files selection" box with 3 buttons "Copy files", "Move files" and "Remove files").
 
 ### Removed